requirements.txt
python>=3.9.7
cudatoolkit>=11.3.1
pytorch>=1.11.0
recbole>=1.2.0